+++ projects/ngx-piano/src/public-api.ts
@@ -3,6 +3,6 @@
  */
 
 export * from './lib/ngx-piano.module';
-export * from './lib/event/action-type';
+export * from './lib/event/ngx-piano-action-type';
 export * from './lib/event/piano-track-click.directive';
 export * from './lib/piano-tracker';